feat: move ExecError definition to codegen
This allows modules to also access ExecErrors, so they can analyze
return codes, etc.
